fix 600 200 200 error

+++ philo/monitor.c
@@ -11,7 +11,7 @@ int	check_philos(t_philo *philos, int i)
 	time = time_ms(&philos[i].tv_beg);
 	pthread_mutex_unlock(&philos[i].time_protection);
 	gettimeofday(&tv, NULL);
-	if (time_ms(&tv) - time >= (long long)t_to_die)
+	if (time_ms(&tv) - time >= t_to_die)
 	{
 		kill_philos(philos, philos[i].stuff->number_of_philos);
 		gettimeofday(&tv, NULL);
